Extract service provider
